+++ bboxdb-commons/pom.xml
@@ -3,7 +3,7 @@
 
 	<modelVersion>4.0.0</modelVersion>
 	<artifactId>bboxdb-commons</artifactId>
-	
+
 	<parent>
 		<groupId>org.bboxdb</groupId>
 		<artifactId>bboxdb-parent</artifactId>
