+++ bboxdb-commons/src/test/java/org/bboxdb/TestCloseableHelper.java
@@ -15,7 +15,7 @@
  *    limitations under the License. 
  *    
  *******************************************************************************/
-package org.bboxdb.tools;
+package org.bboxdb;
 
 import java.io.Closeable;
 import java.io.IOException;
