+++ bboxdb-commons/src/test/java/org/bboxdb/TestThreadHelper.java
@@ -15,7 +15,7 @@
  *    limitations under the License. 
  *    
  *******************************************************************************/
-package org.bboxdb.tools;
+package org.bboxdb;
 
 import java.util.ArrayList;
 import java.util.List;
