+++ bboxdb-server/pom.xml
@@ -84,11 +84,6 @@
 			<version>2.2</version>
 		</dependency>
 		<dependency>
-			<groupId>org.mockito</groupId>
-			<artifactId>mockito-core</artifactId>
-			<version>2.11.0</version>
-		</dependency>
-		<dependency>
 			<artifactId>je</artifactId>
 			<groupId>com.sleepycat</groupId>
 			<version>7.3.7</version>
